Update drl_util module
* `TruncatedSequenceGenerator`
* `copy_module`
* `polyak_update_module`
* `copy_network`
* `polyak_update_network`
